+++ models/rating.py
@@ -1,7 +1,6 @@
 from pydantic import BaseModel
-from decimal import Decimal
 
 
 class Rating(BaseModel):
-    rate: Decimal
+    rate: float
     count: int
